backend/requirements_clean.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
websockets>=11.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cachetools>=5.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
jinja2>=3.1.0

# 现有项目依赖 (复用)
playwright>=1.40.0
asyncio-mqtt>=0.13.0
dataclasses-json>=0.6.0

# AI和向量依赖
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# 系统资源监控
psutil>=5.9.0

# 数据处理
requests>=2.31.0
aiohttp>=3.9.0
pywinpty>=2.0.14
